+++ surveys/base_admin.py
@@ -8,10 +8,12 @@ from .admin_forms import SurveyFieldForm
 class SurveyFieldAdmin(admin.ModelAdmin):
     list_display = ('name',)
     form = SurveyFieldForm
+    list_filter = ('fieldsets',)
 
 
 class SurveyFieldOrderingInline(OrderableTabularInline):
     model = models.SurveyFieldOrdering
+    raw_id_fields = ('field',)
     extra = 0
     min_num = 1
 
@@ -19,11 +21,12 @@ class SurveyFieldOrderingInline(OrderableTabularInline):
 class SurveyFieldsetAdmin(admin.ModelAdmin):
     inlines = [SurveyFieldOrderingInline]
     list_display = ('name',)
+    list_filter = ('surveys',)
 
 
 class SurveyFieldOrderingAdmin(OrderableAdmin):
     list_display = ('field', 'fieldset', 'sort_order_display')
-    list_filter = ('fieldset',)
+    list_filter = ('fieldset', 'field')
 
 
 class SurveyFieldsetOrderingInline(OrderableTabularInline):
@@ -39,8 +42,10 @@ class SurveyAdmin(admin.ModelAdmin):
 
 class SurveyFieldsetOrderingAdmin(OrderableAdmin):
     list_display = ('survey', 'fieldset', 'sort_order_display')
-    list_filter = ('survey',)
+    list_filter = ('survey', 'fieldset')
 
 
 class UserResponseAdmin(OrderableAdmin):
-    list_display = ('fieldset', 'user_id')
+    list_display = ('survey', 'fieldset', 'user_id')
+    list_filter = ('survey', 'fieldset')
+    list_display_links = list_display