+++ web/app/scripts/src/services/fieldsets-parser.es.js
@@ -50,6 +50,7 @@ module.service(moduleProperties.componentName, [
                         type: field.field_type,
                         templateOptions: {
                             fieldSetId: fieldset.id,
+                            autoId: `id-${fieldset.id}-${field.id}`,
                             choices: field.answers,
                             fieldOptions: {
                                 // jscs:disable disallowQuotedKeysInObjects